OGS log file parser: Separate parser from regexes
splitted
