+++ tests/log_parser_parallel.py
@@ -1,6 +1,8 @@
 import unittest
 
 from ogs6py.log_parser.log_parser import parse_file
+# this needs to be replaced with regexes from specific ogs version
+from ogs6py.log_parser.ogs_regexes import ogs_regexes
 import pandas as pd
 from collections import namedtuple, defaultdict
 
@@ -14,6 +16,21 @@ def filter_time_analysis_by_time_step(df):
     return dfe
 
 
+def pandas_from_records(records):
+    df = pd.DataFrame(records)
+    # Some logs do not contain information about time_step and iteration
+    # These information must be collected by context (by surrounding log lines from same mpi_process)
+    # Logs are grouped by mpi_process to get only surrounding log lines from same mpi_process
+
+    # There are log lines that give the current time step (when time step starts).
+    # It can be assumed that in all following lines belong to this time steps, until next collected value of time step
+    df['time_step'] = df.groupby('mpi_process')[['time_step']].fillna(method='ffill').fillna(value=0)
+
+    # Back fill, because iteration number can be found in logs at the END of the iteration
+    df['iteration_number'] = df.groupby('mpi_process')[['iteration_number']].fillna(method='bfill')
+    return df
+
+
 def log_types(records):
     d = defaultdict(list)
     for i in records:
@@ -24,11 +41,11 @@ def log_types(records):
 class OGSParserTest(unittest.TestCase):
     def test_parallel_1_compare_serial_info(self):
         filename_p = 'parser/parallel_1_info.txt'
-        records_p = parse_file(filename_p, petsc=True)
+        records_p = parse_file(filename_p, ogs_regexes(), petsc=True)
         num_of_record_type_p = [len(i) for i in log_types(records_p).values()]
 
         filename_s = 'parser/serial_info.txt'
-        records_s = parse_file(filename_s, petsc=False)
+        records_s = parse_file(filename_s, ogs_regexes(), petsc=False)
         num_of_record_type_s = [len(i) for i in log_types(records_s).values()]
 
         self.assertSequenceEqual(num_of_record_type_s, num_of_record_type_p,
@@ -36,7 +53,7 @@ class OGSParserTest(unittest.TestCase):
 
     def test_parallel_3_debug(self):
         filename = 'parser/parallel_3_debug.txt'
-        records = parse_file(filename, petsc=True)
+        records = parse_file(filename, ogs_regexes(), petsc=True)
         mpi_processes = 3
 
         self.assertEqual(len(records) % mpi_processes, 0,
@@ -46,9 +63,7 @@ class OGSParserTest(unittest.TestCase):
         self.assertEqual(all(i % mpi_processes == 0 for i in num_of_record_type), True,
                          'The number of logs of each type should be a multiple of the number of processes')
 
-        df = pd.DataFrame(records)
-        df['time_step'] = df.groupby('mpi_process')[['time_step']].fillna(method='ffill').fillna(value=0)
-        df['iteration_number'] = df.groupby('mpi_process')[['iteration_number']].fillna(method='bfill')
+        df = pandas_from_records(records)
         dfe = filter_time_analysis_by_time_step(df)
 
         # some specific values
@@ -62,6 +77,21 @@ class OGSParserTest(unittest.TestCase):
         self.assertAlmostEqual(dfe.at[record_id(mpi_process=1.0, time_step=1.0), 'dirichlet_time'], 0.000250, digits)
         self.assertAlmostEqual(dfe.at[record_id(mpi_process=2.0, time_step=1.0), 'time_step_solution_time'], 0.008504,
                                digits)
+
+    def test_seriel_convergence_short(self):
+        filename = 'parser/serial_convergence_short.txt'
+        records = parse_file(filename, ogs_regexes(), petsc=False)
+        df = pandas_from_records(records)
+        # component-wise
+        dfe_sic = df.pivot_table(['dx', 'x', 'dx_x'], ['time_step', 'iteration_number', 'component'])
+        # iteration - wise
+        dfe_si = df.pivot_table(['dx', 'x', 'dx_x'], ['time_step', 'iteration_number'])
+
+        # some specific values
+        record_id = namedtuple('id', 'time_step iteration_number component')
+        digits = 6
+        self.assertAlmostEqual(dfe_sic.at[record_id(time_step=1.0, iteration_number=1, component=1), 'dx'],
+                               9.694038e-02, digits)
 
 
 if __name__ == '__main__':
